+++ gnames.go
@@ -2,14 +2,18 @@ package gnames
 
 import (
 	"context"
+	"unicode"
 
 	"github.com/gnames/gnames/config"
 	"github.com/gnames/gnames/ent/score"
 	"github.com/gnames/gnames/ent/verifier"
 	"github.com/gnames/gnames/io/matcher"
 	"github.com/gnames/gnlib/ent/gnvers"
+	mlib "github.com/gnames/gnlib/ent/matcher"
 	vlib "github.com/gnames/gnlib/ent/verifier"
 	"github.com/gnames/gnmatcher"
+	"github.com/gnames/gnparser/ent/str"
+	"github.com/gnames/gnuuid"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -42,7 +46,17 @@ func (g gnames) Verify(
 	log.Printf("Verifying %d name-strings.", len(params.NameStrings))
 	res := make([]*vlib.Verification, len(params.NameStrings))
 
-	matches := g.matcher.MatchNames(params.NameStrings)
+	var matches []mlib.Match
+
+	if params.WithCapitalization {
+		names := make([]string, len(params.NameStrings))
+		for i := range params.NameStrings {
+			names[i] = str.CapitalizeName(params.NameStrings[i])
+		}
+		matches = g.matcher.MatchNames(names)
+	} else {
+		matches = g.matcher.MatchNames(params.NameStrings)
+	}
 
 	var errString string
 	matchRecords, err := g.vf.MatchRecords(ctx, matches)
@@ -64,6 +78,11 @@ func (g gnames) Verify(
 				PreferredResults: s.PreferredResults(params.PreferredSources, mr),
 				Error:            errString,
 			}
+			if params.WithCapitalization {
+				item.InputCapitalized = true
+				item.Input = params.NameStrings[i]
+				item.InputID = gnuuid.New(item.Input).String()
+			}
 
 			res[i] = &item
 		} else {
@@ -75,4 +94,22 @@ func (g gnames) Verify(
 
 func (g gnames) DataSources(ids ...int) ([]*vlib.DataSource, error) {
 	return g.vf.DataSources(ids...)
+}
+
+func FirstUpperCase(name string) string {
+	runes := []rune(name)
+	if len(runes) < 2 {
+		return name
+	}
+
+	one := runes[0]
+	two := runes[1]
+	if unicode.IsUpper(one) || !unicode.IsLetter(one) {
+		return name
+	}
+	if one == 'x' && (two == ' ' || unicode.IsUpper(two)) {
+		return name
+	}
+	runes[0] = unicode.ToUpper(one)
+	return string(runes)
 }