+++ io/verifierpg/match_records.go
@@ -41,9 +41,9 @@ type verif struct {
 	AcceptedName        sql.NullString
 	Classification      sql.NullString
 	ClassificationRanks sql.NullString
+	ClassificationIds   sql.NullString
 	ParseQuality        int
 }
-
 type matchSplit struct {
 	noMatch   []*mlib.Match
 	canonical []*mlib.Match
@@ -52,7 +52,8 @@ type matchSplit struct {
 var namesQ = `
   SELECT canonical_id, name, data_source_id, record_id, name_string_id,
       local_id, outlink_id, accepted_record_id, accepted_name_id,
-      accepted_name, classification, classification_ranks, parse_quality
+      accepted_name, classification, classification_ranks,
+      classification_ids, parse_quality
     FROM verification where %s in (%s)`
 
 // MatchRecords takes matches from gnmatcher and returns back data from
@@ -241,6 +242,7 @@ func (dgp *verifierpg) populateMatchRecord(
 			IsSynonym:              verifRec.RecordID != verifRec.AcceptedRecordID,
 			ClassificationPath:     verifRec.Classification.String,
 			ClassificationRanks:    verifRec.ClassificationRanks.String,
+			ClassificationIDs:      verifRec.ClassificationIds.String,
 			EditDistance:           edDist,
 			StemEditDistance:       edDistStem,
 			MatchType:              m.MatchType,